refactor: filter functions in utils
